# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aries Next API
  description: Aries Next 博客系统 API（由 utoipa 注解自动生成，权威契约即本文件）
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
servers:
- url: /
paths:
...
  /api/admin/auth/login:
    post:
      tags:
      - Admin Auth
      summary: 管理员登录
      description: 用户名或邮箱 + 密码认证；成功后撤销旧 Session、签发新 Session Cookie（HttpOnly，Path 限定 /api/admin）。账号维度 5 次/15 分钟、IP 维度 30 次/15 分钟双重限流，未命中用户也执行一次 Argon2id 以掩盖账号枚举的时间差异。
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: 认证成功并签发 Session Cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          description: 用户名或密码错误，或账号已被禁用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Origin 校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 触发限流
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/auth/logout:
    post:
      tags:
      - Admin Auth
      summary: 管理员登出
      description: 撤销当前 Session 并清除 Session Cookie，返回 204。
      operationId: logout
      responses:
        '204':
          description: 撤销当前 Session 并清除 Cookie
        '401':
          description: 未认证或 Session 已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Origin 校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
  /api/admin/auth/password/forgot:
    post:
      tags:
      - Admin Auth
      summary: 请求密码重置
      description: 无论账号是否存在都返回相同响应（202），避免账号枚举；每邮箱 3 次/15 分钟限流。Phase 05 接入 Email Adapter 前不会发送 Email，Reset Token 不落库明文。
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
        required: true
      responses:
        '202':
          description: 无论账号是否存在都返回相同响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: 邮箱格式不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Origin 校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 触发限流
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/auth/password/reset:
    post:
      tags:
      - Admin Auth
      summary: 重置密码
      description: 凭 Password Reset Token 设置新密码，成功后撤销该 User 的全部 Session，返回 204；Token 无效或已过期时返回 400。每 Token 维度 3 次/15 分钟限流。
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
        required: true
      responses:
        '204':
          description: 更新 Password 并撤销该 User 的全部 Session
        '400':
          description: Token 无效或已过期，或新密码不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Origin 校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 触发限流
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/auth/session:
    get:
      tags:
      - Admin Auth
      summary: 查询当前会话
      description: 返回当前 Session 对应的 User 视图与 Session 过期时间；无有效 Session Cookie 时返回 401。
      operationId: getSession
      responses:
        '200':
          description: 当前 Session 与 User 视图
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          description: 未认证或 Session 已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
  /api/admin/bootstrap:
    post:
      tags:
      - Admin Auth
      summary: 创建首个 Owner 并完成初始化
      description: 校验一次性 bootstrap_secret 后创建首个 Owner User，并签发 Session Cookie。仅限未初始化时调用，限流 5 次/15 分钟。
      operationId: bootstrap
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
        required: true
      responses:
        '200':
          description: 创建首个 Owner 并签发 Session Cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: 用户名、邮箱、显示名或密码不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: bootstrap_secret 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 系统已初始化
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 触发限流
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/bootstrap/status:
    get:
      tags:
      - Admin Auth
      summary: 查询系统初始化状态
      description: 返回是否已创建首个 User；未初始化时前端应引导进入 Bootstrap 流程。
      operationId: bootstrapStatus
      responses:
        '200':
          description: 是否已创建首个 User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapStatusResponse'
        '500':
          description: 数据库查询失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    BootstrapRequest:
      type: object
      required:
      - bootstrap_secret
      - username
      - email
      - display_name
      - password
      properties:
        bootstrap_secret:
          type: string
          description: 一次性引导密钥（环境变量 BOOTSTRAP_SECRET，≥ 24 字符）。
        display_name:
          type: string
          description: 显示名：1–60 字符。
        email:
          type: string
          description: 邮箱地址。
        password:
          type: string
          description: 初始密码：10–128 字符。
        username:
          type: string
          description: 登录用户名：3–30 位字母、数字、下划线或短横线。
    BootstrapStatusResponse:
      type: object
      required:
      - initialized
      properties:
        initialized:
          type: boolean
          description: 是否已创建首个 User。
...
    ErrorBody:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 机器可读错误码，如 `UNAUTHORIZED`、`ARTICLE_NOT_FOUND`。
        details:
          description: 结构化补充信息（如引用计数），大多数错误为空，序列化时省略。
        message:
          type: string
          description: 人类可读错误信息。
    ErrorResponse:
      type: object
      description: |-
        统一错误响应体（与 `http::error::ApiError` 的实际 JSON 形状一致）：
        `{ "error": { "code", "message", "details?" } }`。
        所有端点的 4xx/5xx 响应统一引用本类型。
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
    ForgotPasswordRequest:
      type: object
      required:
      - email
      properties:
        email:
          type: string
          description: 邮箱地址。
...
    LoginRequest:
      type: object
      required:
      - login
      - password
      properties:
        login:
          type: string
          description: 用户名或邮箱。
        password:
          type: string
          description: 密码。
...
    MessageResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          description: 人类可读结果信息。
...
    ResetPasswordRequest:
      type: object
      required:
      - token
      - password
      properties:
        password:
          type: string
          description: 新密码：10–128 字符。
        token:
          type: string
          description: Password Reset Token。
...
    SessionResponse:
      type: object
      required:
      - user
      - expires_at
      properties:
        expires_at:
          type: string
          format: date-time
          description: Session 过期时间（UTC）。
        user:
          $ref: '#/components/schemas/UserResponse'
          description: 当前 User 视图。
...
    UserResponse:
      type: object
      required:
      - id
      - username
      - email
      - display_name
      - role
      - permissions
      properties:
        avatar_url:
          type:
          - string
          - 'null'
          description: 头像 URL，未设置时为 null。
        display_name:
          type: string
          description: 显示名。
        email:
          type: string
          description: 邮箱地址。
        id:
          type: integer
          format: int64
          description: User ID。
        permissions:
          type: array
          items:
            type: string
          description: 该角色拥有的权限标识列表。
        role:
          type: string
          description: 角色：owner / editor / moderator。
        username:
          type: string
          description: 登录用户名。
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aries_admin_session